Update README.md
Please solve the following problems for Homework 4 at the end of this page. Please read the following notes before attempting to solve.

Notes:

Chapter 5, problem 8: Even though the problem is implicitly asking not use NOT gates, I am instructing the graders to be lenient with that and accept it as a correct answer as long as you do not implement using AND/OR gates.
Chapter 5, problem 16 (c) wants you to implement the two functions using a 4:16 decoder and 2 8-input OR gates. The 4:16 decoder has to be built out of a number of 2:4 decoders (show the connections of building it).
 HWK 4 Summer 2021.pdf   Problems marked in red are bonus problems.